requirements.txt
pandas
sentence-transformers
pinecone-client[grpc]
python-dotenv
